+++ packages/demo/view-libs/react.next@19.x/app/test/page.tsx
@@ -1,75 +1,87 @@
 'use client';
 
-export const rexaStatus = () => {
-	const {store,setState,mutate} = createReaxable({
-		pending: false as boolean|string|number ,
-		error : false as boolean|string|number,
-	});
-	return {
-		status:store,
-		setStatus(status:Partial<typeof store>){
-			setState(status);
-		}
-	}
-};
-
-export const reaxel_UserBio = reaxel(() => {
-	const { status,setStatus } = rexaStatus();
-	const { store , setState , mutate } = createReaxable({
-		bio : null as string ,
-		status ,
-	});
-	
-	const fetchBio = () => {
-		if( store.status.pending ) {
-			return;
-		}
-		new Promise<string>(( resolve , reject ) => {
-			setStatus({pending : true});
-			setTimeout(() => {
-				//Choose which code to comment out to switch different promise results.
-				resolve(`I'm an software engineer.`);
-				// reject('403 Forbidden');
-			} , 2000);
-		}).
-		then(( bio ) => {
-			setState({ bio });
-			setStatus({ pending : false , error : false });
-		}).catch(e => {
-			setStatus({error : e,pending : false})
-		});
-	};
-	const rtn = {
-		UserBio_Store : store ,
-		UserBio_SetState : setState ,
-		UserBio_Mutate : mutate ,
-		fetchBio ,
-	};
-	return () => rtn;
-});
-
-
-
-
 export default reaxper(() => {
-	const {fetchBio,UserBio_Store} = reaxel_UserBio();
-	
-	React.useEffect(() => {
-		fetchBio();
-	} , []);
-	
-	if(UserBio_Store.status.error){
-		return <div>Error Code: {UserBio_Store.status.error}</div>
+	const { statics , setInputName , token , login } = reaxel_Auth();
+	if( token ) {
+		return <Profile />;
+	} else {
+		return <Login />;
 	}
-	if(UserBio_Store.status.pending){
-		return <div>pending...</div>
-	}
-	
-	return <div>{ UserBio_Store.bio }</div>;
 });
 
+const Login = reaxper(( props ) => {
+	const { login , setInputName , statics , logout , logginStatus } = reaxel_Auth();
+	if(logginStatus.pending){
+		return <p>logging...</p>
+	}
+	
+	return <div>
+		<select
+			onChange = { ( e ) => {
+				setInputName(e.target.value);
+			} }
+			value = { reaxel_Auth.store.input_username }
+		>
+			<option
+				value = ""
+				disabled
+				hidden
+			>
+				请选择要登录的用户
+			</option>
+			{ statics.users.map(( { userName } ) => {
+				return <option
+					key = { userName }
+					value = { userName }
+				>{ userName }</option>;
+			}) }
+		</select>
+		<button
+			onClick = { () => {
+				if( !reaxel_Auth.store.input_username ) {
+					confirm('必须选择一个要登录的用户!');
+					return;
+				}
+				login();
+			} }
+		>
+			login
+		</button>
+		{logginStatus.error && <p>
+			{ logginStatus.error.message }
+		</p>}
+		{ reaxel_Auth.store.logged_in && <button
+			onClick = { () => logout() }
+		>
+			logout
+		</button> }
+	</div>;
+});
 
-import { createReaxable } from 'reaxes';
-import * as utils from 'reaxes-utils';
+const Profile = reaxper(( props ) => {
+	const { bio , bioStatus , clearBio } = reaxel_Profile();
+	const { logout } = reaxel_Auth();
+	
+	if( bioStatus.error ) {
+		return <div>Error Code: { bioStatus.error }</div>;
+	}
+	if( bioStatus.pending ) {
+		return <div>requesting user profile...</div>;
+	}
+	
+	return <div>
+		<p>{ bio }</p>
+		<p>
+			<button
+				onClick={() => {
+					logout()
+				}}
+			>logout</button>
+		</p>
+	</div>;
+});
+
+import { reaxel_Auth } from './reaxels/auth';
+import { reaxel_Profile } from './reaxels/user-bio';
 import React from 'react';
-import { reaxel , reaxper } from 'reaxes-react';
+import { reaxper } from 'reaxes-react';